backend/requirements.txt
fastapi
uvicorn
python-dotenv
gitpython
sentence-transformers
supabase
groq